Thêm hàm main và các phương thức tạo File

+++ src/Generator.java
@@ -1,20 +1,23 @@
 import java.io.*;
+import java.nio.file.FileAlreadyExistsException;
 import java.util.Random;
-
-import static sun.java2d.cmm.ColorTransform.Out;
 
 public class Generator {
     public Generator() {
     }
 
-    public void randomNumber(String filePath, int amount) throws IOException {
+    public void randomNumber(String filePath, int amount, int range) throws IOException {
         File text = new File(filePath);
+        if (text.exists()) {
+            throw new FileAlreadyExistsException(filePath);
+        }
+
         FileWriter fileWriter = new FileWriter(text);
         BufferedWriter bufferedWriter = new BufferedWriter(fileWriter);
         Random random = new Random();
 
         for (int index = 0; index < amount; index++) {
-            int numberToWrite = random.nextInt(10);
+            int numberToWrite = random.nextInt(range);
             bufferedWriter.write(String.valueOf(numberToWrite));
             bufferedWriter.newLine();
         }
